+++ src/Facility.ConformanceApi/Http/ConformanceApiHttpMapping.g.cs
@@ -599,5 +599,132 @@ namespace Facility.ConformanceApi.Http
 					}.Build(),
 				},
 			}.Build();
+
+		public static readonly HttpMethodMapping<MirrorBytesRequestDto, MirrorBytesResponseDto> MirrorBytesMapping =
+			new HttpMethodMapping<MirrorBytesRequestDto, MirrorBytesResponseDto>.Builder
+			{
+				HttpMethod = HttpMethod.Post,
+				Path = "/mirrorBytes",
+				ValidateRequest = request =>
+				{
+					if (request.Content == null)
+						return ServiceResult.Failure(ServiceErrors.CreateRequestFieldRequired("content"));
+					return ServiceResult.Success();
+				},
+				GetRequestHeaders = request =>
+					new Dictionary<string, string?>
+					{
+						["Content-Type"] = request.Type,
+					},
+				SetRequestHeaders = (request, headers) =>
+				{
+					headers.TryGetValue("Content-Type", out var headerType);
+					request.Type = headerType;
+					return request;
+				},
+				RequestBodyType = typeof(byte[]),
+				GetRequestBody = request => request.Content,
+				CreateRequest = body => new MirrorBytesRequestDto { Content = (byte[]?) body },
+				ResponseMappings =
+				{
+					new HttpResponseMapping<MirrorBytesResponseDto>.Builder
+					{
+						StatusCode = (HttpStatusCode) 200,
+						ResponseBodyType = typeof(byte[]),
+						MatchesResponse = response => response.Content != null,
+						GetResponseBody = response => response.Content,
+						CreateResponse = body => new MirrorBytesResponseDto { Content = (byte[]?) body },
+					}.Build(),
+				},
+				GetResponseHeaders = response =>
+					new Dictionary<string, string?>
+					{
+						["Content-Type"] = response.Type,
+					},
+				SetResponseHeaders = (response, headers) =>
+				{
+					headers.TryGetValue("Content-Type", out var headerType);
+					response.Type = headerType;
+					return response;
+				},
+			}.Build();
+
+		public static readonly HttpMethodMapping<MirrorTextRequestDto, MirrorTextResponseDto> MirrorTextMapping =
+			new HttpMethodMapping<MirrorTextRequestDto, MirrorTextResponseDto>.Builder
+			{
+				HttpMethod = HttpMethod.Post,
+				Path = "/mirrorText",
+				ValidateRequest = request =>
+				{
+					if (request.Content == null)
+						return ServiceResult.Failure(ServiceErrors.CreateRequestFieldRequired("content"));
+					return ServiceResult.Success();
+				},
+				GetRequestHeaders = request =>
+					new Dictionary<string, string?>
+					{
+						["Content-Type"] = request.Type,
+					},
+				SetRequestHeaders = (request, headers) =>
+				{
+					headers.TryGetValue("Content-Type", out var headerType);
+					request.Type = headerType;
+					return request;
+				},
+				RequestBodyType = typeof(string),
+				GetRequestBody = request => request.Content,
+				CreateRequest = body => new MirrorTextRequestDto { Content = (string?) body },
+				ResponseMappings =
+				{
+					new HttpResponseMapping<MirrorTextResponseDto>.Builder
+					{
+						StatusCode = (HttpStatusCode) 200,
+						ResponseBodyType = typeof(string),
+						MatchesResponse = response => response.Content != null,
+						GetResponseBody = response => response.Content,
+						CreateResponse = body => new MirrorTextResponseDto { Content = (string?) body },
+					}.Build(),
+				},
+				GetResponseHeaders = response =>
+					new Dictionary<string, string?>
+					{
+						["Content-Type"] = response.Type,
+					},
+				SetResponseHeaders = (response, headers) =>
+				{
+					headers.TryGetValue("Content-Type", out var headerType);
+					response.Type = headerType;
+					return response;
+				},
+			}.Build();
+
+		public static readonly HttpMethodMapping<BodyTypesRequestDto, BodyTypesResponseDto> BodyTypesMapping =
+			new HttpMethodMapping<BodyTypesRequestDto, BodyTypesResponseDto>.Builder
+			{
+				HttpMethod = HttpMethod.Post,
+				Path = "/bodyTypes",
+				ValidateRequest = request =>
+				{
+					if (request.Content == null)
+						return ServiceResult.Failure(ServiceErrors.CreateRequestFieldRequired("content"));
+					return ServiceResult.Success();
+				},
+				RequestBodyType = typeof(string),
+				RequestBodyContentType = "text/x-input",
+				GetRequestBody = request => request.Content,
+				CreateRequest = body => new BodyTypesRequestDto { Content = (string?) body },
+				ResponseMappings =
+				{
+					new HttpResponseMapping<BodyTypesResponseDto>.Builder
+					{
+						StatusCode = (HttpStatusCode) 200,
+						ResponseBodyType = typeof(byte[]),
+						ResponseBodyContentType = "application/x-output",
+						MatchesResponse = response => response.Content != null,
+						GetResponseBody = response => response.Content,
+						CreateResponse = body => new BodyTypesResponseDto { Content = (byte[]?) body },
+					}.Build(),
+				},
+			}.Build();
 	}
 }