+++ src/Facility.ConformanceApi/Http/HttpClientConformanceApi.g.cs
@@ -77,5 +77,14 @@ namespace Facility.ConformanceApi.Http
 
 		public Task<ServiceResult<RequiredResponseDto>> RequiredAsync(RequiredRequestDto request, CancellationToken cancellationToken = default) =>
 			TrySendRequestAsync(ConformanceApiHttpMapping.RequiredMapping, request, cancellationToken);
+
+		public Task<ServiceResult<MirrorBytesResponseDto>> MirrorBytesAsync(MirrorBytesRequestDto request, CancellationToken cancellationToken = default) =>
+			TrySendRequestAsync(ConformanceApiHttpMapping.MirrorBytesMapping, request, cancellationToken);
+
+		public Task<ServiceResult<MirrorTextResponseDto>> MirrorTextAsync(MirrorTextRequestDto request, CancellationToken cancellationToken = default) =>
+			TrySendRequestAsync(ConformanceApiHttpMapping.MirrorTextMapping, request, cancellationToken);
+
+		public Task<ServiceResult<BodyTypesResponseDto>> BodyTypesAsync(BodyTypesRequestDto request, CancellationToken cancellationToken = default) =>
+			TrySendRequestAsync(ConformanceApiHttpMapping.BodyTypesMapping, request, cancellationToken);
 	}
 }