+++ src/Facility.Core/Http/HttpClientServiceSettings.cs
@@ -25,6 +25,16 @@ namespace Facility.Core.Http
 		public HttpContentSerializer? ContentSerializer { get; set; }
 
 		/// <summary>
+		/// The content serializer used by requests and responses for bytes (optional).
+		/// </summary>
+		public HttpContentSerializer? BytesSerializer { get; set; }
+
+		/// <summary>
+		/// The content serializer used by requests and responses for text (optional).
+		/// </summary>
+		public HttpContentSerializer? TextSerializer { get; set; }
+
+		/// <summary>
 		/// The aspects used when sending requests and receiving responses (optional).
 		/// </summary>
 		public IReadOnlyList<HttpClientServiceAspect>? Aspects { get; set; }