+++ src/Facility.Core/Http/HttpContentSerializer.cs
@@ -16,20 +16,19 @@ namespace Facility.Core.Http
 		public string DefaultMediaType => DefaultMediaTypeCore;
 
 		/// <summary>
-		/// Determines if the specified media type is supported.
+		/// Determines if the specified media type can be read by this serializer.
 		/// </summary>
 		public bool IsSupportedMediaType(string mediaType) => IsSupportedMediaTypeCore(mediaType);
 
 		/// <summary>
+		/// Determines if the specified media type is accepted when investigating the Accept header.
+		/// </summary>
+		public bool IsAcceptedMediaType(string mediaType) => IsAcceptedMediaTypeCore(mediaType);
+
+		/// <summary>
 		/// Creates HTTP content for the specified DTO.
 		/// </summary>
-		public HttpContent CreateHttpContent(object content, string? mediaType = null)
-		{
-			if (mediaType != null && !IsSupportedMediaType(mediaType))
-				throw new ArgumentException($"Unsupported media type '{mediaType}'.");
-
-			return CreateHttpContentCore(content, mediaType);
-		}
+		public HttpContent CreateHttpContent(object content, string? mediaType = null) => CreateHttpContentCore(content, mediaType);
 
 		/// <summary>
 		/// Reads a DTO from the specified HTTP content.
@@ -41,30 +40,36 @@ namespace Facility.Core.Http
 		}
 
 		/// <summary>
-		/// Reads a DTO from the specified HTTP content.
+		/// Reads an object from the specified HTTP content.
 		/// </summary>
-		public async Task<ServiceResult<object>> ReadHttpContentAsync(Type dtoType, HttpContent? content, CancellationToken cancellationToken = default)
+		public async Task<ServiceResult<object>> ReadHttpContentAsync(Type objectType, HttpContent? content, CancellationToken cancellationToken = default)
 		{
 			var contentType = content?.Headers.ContentType;
 			if (contentType == null)
 				return ServiceResult.Failure(HttpServiceErrors.CreateMissingContentType());
 
-			string mediaType = contentType.MediaType;
+			var mediaType = contentType.MediaType;
 			if (!IsSupportedMediaType(mediaType))
 				return ServiceResult.Failure(HttpServiceErrors.CreateUnsupportedContentType(mediaType));
 
-			return await ReadHttpContentAsyncCore(dtoType, content!, cancellationToken).ConfigureAwait(false);
+			return await ReadHttpContentAsyncCore(objectType, content!, cancellationToken).ConfigureAwait(false);
 		}
 
 		/// <summary>
-		/// The media type for requests.
+		/// The default media type for the serializer.
 		/// </summary>
 		protected abstract string DefaultMediaTypeCore { get; }
 
 		/// <summary>
-		/// Determines if the specified media type is supported.
+		/// Determines if the specified media type can be read by this serializer.
 		/// </summary>
 		protected abstract bool IsSupportedMediaTypeCore(string mediaType);
+
+		/// <summary>
+		/// Determines if the specified media type is accepted when investigating the Accept header.
+		/// </summary>
+		/// <remarks>Calls <see cref="IsSupportedMediaType"/> by default.</remarks>
+		protected virtual bool IsAcceptedMediaTypeCore(string mediaType) => IsSupportedMediaType(mediaType);
 
 		/// <summary>
 		/// Creates HTTP content for the specified DTO.
@@ -74,6 +79,6 @@ namespace Facility.Core.Http
 		/// <summary>
 		/// Reads a DTO from the specified HTTP content.
 		/// </summary>
-		protected abstract Task<ServiceResult<object>> ReadHttpContentAsyncCore(Type dtoType, HttpContent content, CancellationToken cancellationToken);
+		protected abstract Task<ServiceResult<object>> ReadHttpContentAsyncCore(Type objectType, HttpContent content, CancellationToken cancellationToken);
 	}
 }